# api.yml
openapi: 3.0.1
info:
  title: ExtranetAPI
  version: '1.0'
paths:
  /analytics/numerical-indicators:
    post:
      tags:
        - Analytics
      summary: Числовые показатели за период
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NumericalPeriodDto'
          text/json:
            schema:
              $ref: '#/components/schemas/NumericalPeriodDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/NumericalPeriodDto'
      responses:
        '200':
          description: Числовые показатели за период
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/NumericalIndicatorsDto'
            application/json:
              schema:
                $ref: '#/components/schemas/NumericalIndicatorsDto'
            text/json:
              schema:
                $ref: '#/components/schemas/NumericalIndicatorsDto'
  /analytics/data-by-date:
    post:
      tags:
        - Analytics
      summary: Получить данные для аналитики по дате
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChartDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ChartDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ChartDto'
      responses:
        '200':
          description: Получить данные для аналитики по дате
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemDto'
  /analytics/data-by-period:
    post:
      tags:
        - Analytics
      summary: Получить данные для аналитики по периоду
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChartDto'
          text/json:
            schema:
              $ref: '#/components/schemas/ChartDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/ChartDto'
      responses:
        '200':
          description: Получить данные для аналитики по дате
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemWeeksDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemWeeksDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChartItemWeeksDto'
...
  /buisness-costs:
    get:
      tags:
        - BuisnessCost
      summary: Получить все бизнесовые издержки
      responses:
        '200':
          description: Получить все издержки
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BuisnessCost'
    post:
      tags:
        - BuisnessCost
      summary: Создать бизнесовую издержку
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          text/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
        required: true
      responses:
        '200':
          description: Создать издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /buisness-costs/{costId}/cost-updating:
    post:
      tags:
        - BuisnessCost
      summary: Обновить бизнесовую издержку
      parameters:
        - name: costId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          text/json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/BuisnessCost'
        required: true
      responses:
        '200':
          description: Обновить издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /buisness-costs/{id}:
    delete:
      tags:
        - BuisnessCost
      summary: Удалить бизнесовую издержку
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
  /costs/{costId}:
    get:
      tags:
        - Cost
      summary: Получить издержку по идентификатору
      parameters:
        - name: costId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Получить издержку по идентификатору
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Cost'
            application/json:
              schema:
                $ref: '#/components/schemas/Cost'
            text/json:
              schema:
                $ref: '#/components/schemas/Cost'
    delete:
      tags:
        - Cost
      summary: Удалить издержку
      parameters:
        - name: costId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Удалить издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /costs:
    get:
      tags:
        - Cost
      summary: Получить все издержки
      responses:
        '200':
          description: Получить все издержки
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cost'
    post:
      tags:
        - Cost
      summary: Создать издержку
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cost'
          text/json:
            schema:
              $ref: '#/components/schemas/Cost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Cost'
        required: true
      responses:
        '200':
          description: Создать издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
  /costs/{costId}/cost-updating:
    post:
      tags:
        - Cost
      summary: Обновить издержку
      parameters:
        - name: costId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cost'
          text/json:
            schema:
              $ref: '#/components/schemas/Cost'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Cost'
        required: true
      responses:
        '200':
          description: Обновить издержку
          content:
            text/plain:
              schema:
                type: integer
                format: int32
            application/json:
              schema:
                type: integer
                format: int32
            text/json:
              schema:
                type: integer
                format: int32
...
components:
  schemas:
...
    BuisnessCost:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        amount:
          type: number
          format: double
        date:
          type: string
          format: date-time
      additionalProperties: false
    ChartDto:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/Period'
        chartPeriodType:
          $ref: '#/components/schemas/ChartPeriodType'
        chartType:
          $ref: '#/components/schemas/ChartType'
      additionalProperties: false
    ChartItemDto:
      type: object
      properties:
        date:
          type: string
          format: date-time
        value:
          type: number
          format: double
      additionalProperties: false
    ChartItemWeeksDto:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/Period'
        value:
          type: number
          format: double
      additionalProperties: false
    ChartPeriodType:
      enum:
        - 0
        - 1
        - 2
        - 3
      type: integer
      format: int32
    ChartType:
      enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
      type: integer
      format: int32
...
    Cost:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
      additionalProperties: false
...
    NumericalIndicatorsDto:
      type: object
      properties:
        averageCheck:
          type: number
          format: double
        averageProductionDays:
          type: integer
          format: int32
        numberOfProducts:
          type: integer
          format: int32
      additionalProperties: false
    NumericalPeriodDto:
      type: object
      properties:
        period:
          $ref: '#/components/schemas/Period'
        chartPeriodType:
          $ref: '#/components/schemas/ChartPeriodType'
      additionalProperties: false
    Period:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      additionalProperties: false
...