# back/FurnitureHelper/src/ExtranetAPI/Models/Period.cs
namespace ExtranetAPI.Models;

public class Period
{
    public Period(DateTime startDate, DateTime endDate)
    {
        StartDate = startDate;
        EndDate = endDate;
    }

    public DateTime StartDate { get; set; }
    public DateTime EndDate { get; set; }
}